Changed all long double to floats

+++ Positioning/LatitudeDegree.cpp
@@ -5,11 +5,11 @@ using namespace std;
 
 LatitudeDegree::LatitudeDegree(const std::string &koordString, char direction) {
     int deg = 0;
-    long double min = 0;
+    float min = 0;
     stringstream(koordString.substr(0, 2)) >> deg;
     stringstream(koordString.substr(2, 10)) >> min;
 
-    long double jdeg = deg + (min / 60);
+    float jdeg = deg + (min / 60);
     if (direction == 'S') {
         jdeg = -jdeg;
     }
@@ -22,12 +22,12 @@ LatitudeDegree operator-(const LatitudeDegree &dg1, const LatitudeDegree &dg2) {
 }
 
 //Returns Distance to Point Zero
-long double LatitudeDegree::toMeters() const {
+float LatitudeDegree::toMeters() const {
     return std::abs(degrees) * 60 * 1852.216;
 }
 
 
-LatitudeDegree::LatitudeDegree(long double degrees) {
+LatitudeDegree::LatitudeDegree(float degrees) {
     if (degrees < -90 || degrees > 90) {
         return;
     }
